Update ALFA-K.R
I discovered a potential issue in ALFA-K that might lead to an error during execution. Specifically, in the ALFA-K.R file at line 210, the parameter u in the dbinom function needs to be explicitly converted to numeric using as.numeric(u). Otherwise, the following error occurs: “Non-numeric argument to mathematical function”.